[feature] Basic login
The login flow is not working properly but I wanted to have something to start.
In the process of writing it, I had to learn how channels work and how to setup
them for client/server communication.  Using atom is pretty awesome, but I can
see how performing mutation everywhere can become a problem.  I will need so
sort of separation between mutation and pure functions.

Now that I know how the communication with the server works, I focus on how the
game works in the client-side and then connect the wires with the server-side.
I believe that the repl-programming style will help me to get really far before
implementing the interaction with the server.